Remove Qwen3Next from model scope
