+++ examples/auto_deploy/model_registry/configs/qwen3Next.yaml
@@ -7,9 +7,8 @@ max_batch_size: 64
 enable_chunked_prefill: true
 model_factory: AutoModelForCausalLM
 # disable_overlap_scheduler: false
-cuda_graph_config:
-  max_batch_size: 64
-  batch_sizes: [1, 2, 4, 8, 16, 32, 64]
+# cuda_graph_config:
+#   batch_sizes: [1, 2, 4, 8, 16, 32, 64]
 kv_cache_config:
   enable_block_reuse: false
   free_gpu_memory_fraction: 0.7
