+++ examples/auto_deploy/model_registry/models.yaml
@@ -373,7 +373,7 @@ models:
 # --- Qwen3-Next (Sep 2025) ---
 - name: Qwen/Qwen3-Next-80B-A3B-Instruct
   config_id: qwen3next
-  yaml_extra: ['world_size_4.yaml', 'qwen3Next.yaml']
+  yaml_extra: ['qwen3Next.yaml']
 # --- OLMo 3 (Nov 2025) ---
 - name: allenai/Olmo-3-7B-Instruct
   config_id: default_ws_2
